+++ cocos/editor-support/IOBuffer.cpp
@@ -23,23 +23,23 @@
  ****************************************************************************/
 #include "IOBuffer.h"
 
-namespace editor {
+MIDDLEWARE_BEGIN
 
-    void IOBuffer::resize (std::size_t newLen, bool needCopy)
+void IOBuffer::resize (std::size_t newLen, bool needCopy)
+{
+    if (_bufferSize < newLen)
     {
-        if (_bufferSize < newLen)
+        uint8_t* newBuffer = new uint8_t[newLen];
+        if (needCopy)
         {
-            uint8_t* newBuffer = new uint8_t[newLen];
-            if (needCopy)
-            {
-                memcpy(newBuffer, _buffer, _bufferSize);
-            }
-            
-            delete[] _buffer;
-            _buffer = newBuffer;
-            _bufferSize = newLen;
-            _outRange = false;
+            memcpy(newBuffer, _buffer, _bufferSize);
         }
+        
+        delete[] _buffer;
+        _buffer = newBuffer;
+        _bufferSize = newLen;
+        _outRange = false;
     }
-
 }
+
+MIDDLEWARE_END
