+++ cocos/editor-support/IOBuffer.h
@@ -26,172 +26,174 @@
 #include <cstddef>
 #include <cstdint>
 #include <string>
+#include "Macro.h"
 
-namespace editor {
-    /**
-     * Use for such as vertex buffer or index buffer,write bytes in it.
-     * If write data out of range,will sign it in _outRange property.
-     * User decide to enlarge it or not.
-     */
-    class IOBuffer
+MIDDLEWARE_BEGIN
+/**
+ * Use for such as vertex buffer or index buffer,write bytes in it.
+ * If write data out of range,will sign it in _outRange property.
+ * User decide to enlarge it or not.
+ */
+class IOBuffer
+{
+public:
+    IOBuffer (std::size_t defaultSize)
     {
-    public:
-        IOBuffer (std::size_t defaultSize)
-        {
-            _bufferSize = defaultSize;
-            _buffer = new uint8_t[_bufferSize];
-        }
-        
-        IOBuffer () {}
-        
-        virtual ~IOBuffer ()
-        {
-            if (_buffer)
-            {
-                delete[] _buffer;
-                _buffer = nullptr;
-            }
-        }
-        
-        inline void writeUint32 (std::size_t pos, uint32_t val)
-        {
-            if (_bufferSize < pos + sizeof(val))
-            {
-                _outRange = true;
-                return;
-            }
-            uint32_t* buffer = (uint32_t*)(_buffer + pos);
-            *buffer = val;
-        }
-        
-        inline void writeFloat32 (std::size_t pos, float val)
-        {
-            if (_bufferSize < pos + sizeof(val))
-            {
-                _outRange = true;
-                return;
-            }
-            float* buffer = (float*)(_buffer + pos);
-            *buffer = val;
-        }
-        
-        inline void writeBytes (const char* bytes, std::size_t bytesLen)
-        {
-            if (_bufferSize < _curPos + bytesLen)
-            {
-                _outRange = true;
-                return;
-            }
-            memcpy(_buffer + _curPos, bytes, bytesLen);
-            _curPos += bytesLen;
-        }
-        
-        inline void writeUint32 (uint32_t val)
-        {
-            if (_bufferSize < _curPos + sizeof(val))
-            {
-                _outRange = true;
-                return;
-            }
-            uint32_t* buffer = (uint32_t*)(_buffer + _curPos);
-            *buffer = val;
-            _curPos += sizeof(val);
-        }
-        
-        inline void writeFloat32 (float val)
-        {
-            if (_bufferSize < _curPos + sizeof(val))
-            {
-                _outRange = true;
-                return;
-            }
-            float* buffer = (float*)(_buffer + _curPos);
-            *buffer = val;
-            _curPos += sizeof(val);
-        }
-        
-        inline void writeUint16 (uint16_t val)
-        {
-            if (_bufferSize < _curPos + sizeof(val))
-            {
-                _outRange = true;
-                return;
-            }
-            uint16_t* buffer = (uint16_t*)(_buffer + _curPos);
-            *buffer = val;
-            _curPos += sizeof(val);
-        }
-        
-        inline uint32_t readUint32 ()
-        {
-            uint32_t* buffer = (uint32_t*)(_buffer + _readPos);
-            _readPos += sizeof(uint32_t);
-            return *buffer;
-        }
-        
-        inline uint16_t readUint16 ()
-        {
-            uint16_t* buffer = (uint16_t*)(_buffer + _readPos);
-            _readPos += sizeof(uint16_t);
-            return *buffer;
-        }
-        
-        inline float readFloat32 ()
-        {
-            float* buffer = (float*)(_buffer + _readPos);
-            _readPos += sizeof(float);
-            return *buffer;
-        }
-        
-        inline char readUint8 ()
-        {
-            char* buffer = (char*)(_buffer + _readPos);
-            _readPos += sizeof(char);
-            return *buffer;
-        }
-        
-        inline void reset ()
-        {
-            _curPos = 0;
-            _readPos = 0;
-        }
-        
-        inline std::size_t length() const
-        {
-            return _curPos;
-        }
-
-        inline std::size_t getCurPos() const
-        {
-            return _curPos;
-        }
-
-        inline uint8_t* getBuffer() const
-        {
-            return _buffer;
-        }
-        
-        inline std::size_t getCapacity() const
-        {
-            return _bufferSize;
-        }
-        
-        inline bool isOutRange()
-        {
-            return _outRange;
-        }
+        _bufferSize = defaultSize;
+        _buffer = new uint8_t[_bufferSize];
+    }
     
-        /**
-         * @brief Resize buffer
-         * @param[in] newLen New size you want to adjustment.
-         * @param[in] needCopy If true,will copy old data to new buffer,default false.
-         */
-        virtual void resize(std::size_t newLen, bool needCopy = false);
-    protected:
-        uint8_t*                    _buffer = nullptr;
-        std::size_t                 _bufferSize = 0;
-        std::size_t                 _curPos = 0;
-        std::size_t                 _readPos = 0;
-        bool                        _outRange = false;
-    };
-}
+    IOBuffer () {}
+    
+    virtual ~IOBuffer ()
+    {
+        if (_buffer)
+        {
+            delete[] _buffer;
+            _buffer = nullptr;
+        }
+    }
+    
+    inline void writeUint32 (std::size_t pos, uint32_t val)
+    {
+        if (_bufferSize < pos + sizeof(val))
+        {
+            _outRange = true;
+            return;
+        }
+        uint32_t* buffer = (uint32_t*)(_buffer + pos);
+        *buffer = val;
+    }
+    
+    inline void writeFloat32 (std::size_t pos, float val)
+    {
+        if (_bufferSize < pos + sizeof(val))
+        {
+            _outRange = true;
+            return;
+        }
+        float* buffer = (float*)(_buffer + pos);
+        *buffer = val;
+    }
+    
+    inline void writeBytes (const char* bytes, std::size_t bytesLen)
+    {
+        if (_bufferSize < _curPos + bytesLen)
+        {
+            _outRange = true;
+            return;
+        }
+        memcpy(_buffer + _curPos, bytes, bytesLen);
+        _curPos += bytesLen;
+    }
+    
+    inline void writeUint32 (uint32_t val)
+    {
+        if (_bufferSize < _curPos + sizeof(val))
+        {
+            _outRange = true;
+            return;
+        }
+        uint32_t* buffer = (uint32_t*)(_buffer + _curPos);
+        *buffer = val;
+        _curPos += sizeof(val);
+    }
+    
+    inline void writeFloat32 (float val)
+    {
+        if (_bufferSize < _curPos + sizeof(val))
+        {
+            _outRange = true;
+            return;
+        }
+        float* buffer = (float*)(_buffer + _curPos);
+        *buffer = val;
+        _curPos += sizeof(val);
+    }
+    
+    inline void writeUint16 (uint16_t val)
+    {
+        if (_bufferSize < _curPos + sizeof(val))
+        {
+            _outRange = true;
+            return;
+        }
+        uint16_t* buffer = (uint16_t*)(_buffer + _curPos);
+        *buffer = val;
+        _curPos += sizeof(val);
+    }
+    
+    inline uint32_t readUint32 ()
+    {
+        uint32_t* buffer = (uint32_t*)(_buffer + _readPos);
+        _readPos += sizeof(uint32_t);
+        return *buffer;
+    }
+    
+    inline uint16_t readUint16 ()
+    {
+        uint16_t* buffer = (uint16_t*)(_buffer + _readPos);
+        _readPos += sizeof(uint16_t);
+        return *buffer;
+    }
+    
+    inline float readFloat32 ()
+    {
+        float* buffer = (float*)(_buffer + _readPos);
+        _readPos += sizeof(float);
+        return *buffer;
+    }
+    
+    inline char readUint8 ()
+    {
+        char* buffer = (char*)(_buffer + _readPos);
+        _readPos += sizeof(char);
+        return *buffer;
+    }
+    
+    inline void reset ()
+    {
+        _curPos = 0;
+        _readPos = 0;
+    }
+    
+    inline std::size_t length() const
+    {
+        return _curPos;
+    }
+
+    inline std::size_t getCurPos() const
+    {
+        return _curPos;
+    }
+
+    inline uint8_t* getBuffer() const
+    {
+        return _buffer;
+    }
+    
+    inline std::size_t getCapacity() const
+    {
+        return _bufferSize;
+    }
+    
+    inline bool isOutRange()
+    {
+        return _outRange;
+    }
+
+    /**
+     * @brief Resize buffer
+     * @param[in] newLen New size you want to adjustment.
+     * @param[in] needCopy If true,will copy old data to new buffer,default false.
+     */
+    virtual void resize(std::size_t newLen, bool needCopy = false);
+protected:
+    uint8_t*                    _buffer = nullptr;
+    std::size_t                 _bufferSize = 0;
+    std::size_t                 _curPos = 0;
+    std::size_t                 _readPos = 0;
+    bool                        _outRange = false;
+};
+
+MIDDLEWARE_END
