+++ cocos/scripting/js-bindings/auto/jsb_cocos2dx_editor_support_auto.cpp
@@ -1,16 +1,16 @@
 #include "scripting/js-bindings/auto/jsb_cocos2dx_editor_support_auto.hpp"
-#if USE_EDITOR_SUPPORT > 0
+#if USE_MIDDLEWARE > 0
 #include "scripting/js-bindings/manual/jsb_conversions.hpp"
 #include "scripting/js-bindings/manual/jsb_global.h"
-#include "editor-support/editor-adapter.h"
-#include "editor-support/EditorManager.h"
+#include "editor-support/middleware-adapter.h"
+#include "editor-support/MiddlewareManager.h"
 
-se::Object* __jsb_editor_Texture2D_proto = nullptr;
-se::Class* __jsb_editor_Texture2D_class = nullptr;
+se::Object* __jsb_cocos2d_middleware_Texture2D_proto = nullptr;
+se::Class* __jsb_cocos2d_middleware_Texture2D_class = nullptr;
 
 static bool js_cocos2dx_editor_support_Texture2D_getRealTextureIndex(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_getRealTextureIndex : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -28,7 +28,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_getRealTextureIndex)
 
 static bool js_cocos2dx_editor_support_Texture2D_setTexParamCallback(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_setTexParamCallback : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -80,7 +80,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_setTexParamCallback)
 
 static bool js_cocos2dx_editor_support_Texture2D_setPixelsHigh(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_setPixelsHigh : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -99,7 +99,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_setPixelsHigh)
 
 static bool js_cocos2dx_editor_support_Texture2D_setPixelsWide(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_setPixelsWide : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -118,7 +118,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_setPixelsWide)
 
 static bool js_cocos2dx_editor_support_Texture2D_getPixelsHigh(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_getPixelsHigh : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -136,7 +136,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_getPixelsHigh)
 
 static bool js_cocos2dx_editor_support_Texture2D_getPixelsWide(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_getPixelsWide : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -154,7 +154,7 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_getPixelsWide)
 
 static bool js_cocos2dx_editor_support_Texture2D_setRealTextureIndex(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_setRealTextureIndex : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
@@ -173,13 +173,13 @@ SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_setRealTextureIndex)
 
 static bool js_cocos2dx_editor_support_Texture2D_setTexParameters(se::State& s)
 {
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_Texture2D_setTexParameters : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 1) {
-        editor::Texture2D::_TexParams arg0;
+        cocos2d::middleware::Texture2D::_TexParams arg0;
         #pragma warning NO CONVERSION TO NATIVE FOR _TexParams
         ok = false;
         SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_Texture2D_setTexParameters : Error processing arguments");
@@ -191,27 +191,27 @@ static bool js_cocos2dx_editor_support_Texture2D_setTexParameters(se::State& s)
 }
 SE_BIND_FUNC(js_cocos2dx_editor_support_Texture2D_setTexParameters)
 
-SE_DECLARE_FINALIZE_FUNC(js_editor_Texture2D_finalize)
+SE_DECLARE_FINALIZE_FUNC(js_cocos2d_middleware_Texture2D_finalize)
 
 static bool js_cocos2dx_editor_support_Texture2D_constructor(se::State& s)
 {
-    editor::Texture2D* cobj = new (std::nothrow) editor::Texture2D();
+    cocos2d::middleware::Texture2D* cobj = new (std::nothrow) cocos2d::middleware::Texture2D();
     s.thisObject()->setPrivateData(cobj);
     return true;
 }
-SE_BIND_CTOR(js_cocos2dx_editor_support_Texture2D_constructor, __jsb_editor_Texture2D_class, js_editor_Texture2D_finalize)
+SE_BIND_CTOR(js_cocos2dx_editor_support_Texture2D_constructor, __jsb_cocos2d_middleware_Texture2D_class, js_cocos2d_middleware_Texture2D_finalize)
 
 
 
 
-static bool js_editor_Texture2D_finalize(se::State& s)
+static bool js_cocos2d_middleware_Texture2D_finalize(se::State& s)
 {
-    CCLOGINFO("jsbindings: finalizing JS object %p (editor::Texture2D)", s.nativeThisObject());
-    editor::Texture2D* cobj = (editor::Texture2D*)s.nativeThisObject();
+    CCLOGINFO("jsbindings: finalizing JS object %p (cocos2d::middleware::Texture2D)", s.nativeThisObject());
+    cocos2d::middleware::Texture2D* cobj = (cocos2d::middleware::Texture2D*)s.nativeThisObject();
     cobj->release();
     return true;
 }
-SE_BIND_FINALIZE_FUNC(js_editor_Texture2D_finalize)
+SE_BIND_FINALIZE_FUNC(js_cocos2d_middleware_Texture2D_finalize)
 
 bool js_register_cocos2dx_editor_support_Texture2D(se::Object* obj)
 {
@@ -225,167 +225,167 @@ bool js_register_cocos2dx_editor_support_Texture2D(se::Object* obj)
     cls->defineFunction("getPixelsWide", _SE(js_cocos2dx_editor_support_Texture2D_getPixelsWide));
     cls->defineFunction("setRealTextureIndex", _SE(js_cocos2dx_editor_support_Texture2D_setRealTextureIndex));
     cls->defineFunction("setTexParameters", _SE(js_cocos2dx_editor_support_Texture2D_setTexParameters));
-    cls->defineFinalizeFunction(_SE(js_editor_Texture2D_finalize));
+    cls->defineFinalizeFunction(_SE(js_cocos2d_middleware_Texture2D_finalize));
     cls->install();
-    JSBClassType::registerClass<editor::Texture2D>(cls);
+    JSBClassType::registerClass<cocos2d::middleware::Texture2D>(cls);
 
-    __jsb_editor_Texture2D_proto = cls->getProto();
-    __jsb_editor_Texture2D_class = cls;
+    __jsb_cocos2d_middleware_Texture2D_proto = cls->getProto();
+    __jsb_cocos2d_middleware_Texture2D_class = cls;
 
     se::ScriptEngine::getInstance()->clearException();
     return true;
 }
 
-se::Object* __jsb_editor_EditorManager_proto = nullptr;
-se::Class* __jsb_editor_EditorManager_class = nullptr;
+se::Object* __jsb_cocos2d_middleware_MiddlewareManager_proto = nullptr;
+se::Class* __jsb_cocos2d_middleware_MiddlewareManager_class = nullptr;
 
-static bool js_cocos2dx_editor_support_EditorManager_addTimer(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_addTimer(se::State& s)
 {
-    editor::EditorManager* cobj = (editor::EditorManager*)s.nativeThisObject();
-    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_EditorManager_addTimer : Invalid Native Object");
+    cocos2d::middleware::MiddlewareManager* cobj = (cocos2d::middleware::MiddlewareManager*)s.nativeThisObject();
+    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_MiddlewareManager_addTimer : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 1) {
-        editor::IEditor* arg0 = nullptr;
+        cocos2d::middleware::IMiddleware* arg0 = nullptr;
         ok &= seval_to_native_ptr(args[0], &arg0);
-        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_EditorManager_addTimer : Error processing arguments");
+        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_MiddlewareManager_addTimer : Error processing arguments");
         cobj->addTimer(arg0);
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 1);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_addTimer)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_addTimer)
 
-static bool js_cocos2dx_editor_support_EditorManager_removeTimer(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_removeTimer(se::State& s)
 {
-    editor::EditorManager* cobj = (editor::EditorManager*)s.nativeThisObject();
-    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_EditorManager_removeTimer : Invalid Native Object");
+    cocos2d::middleware::MiddlewareManager* cobj = (cocos2d::middleware::MiddlewareManager*)s.nativeThisObject();
+    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_MiddlewareManager_removeTimer : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 1) {
-        editor::IEditor* arg0 = nullptr;
+        cocos2d::middleware::IMiddleware* arg0 = nullptr;
         ok &= seval_to_native_ptr(args[0], &arg0);
-        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_EditorManager_removeTimer : Error processing arguments");
+        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_MiddlewareManager_removeTimer : Error processing arguments");
         cobj->removeTimer(arg0);
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 1);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_removeTimer)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_removeTimer)
 
-static bool js_cocos2dx_editor_support_EditorManager_getGLVBID(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_getGLVBID(se::State& s)
 {
-    editor::EditorManager* cobj = (editor::EditorManager*)s.nativeThisObject();
-    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_EditorManager_getGLVBID : Invalid Native Object");
+    cocos2d::middleware::MiddlewareManager* cobj = (cocos2d::middleware::MiddlewareManager*)s.nativeThisObject();
+    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_MiddlewareManager_getGLVBID : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 0) {
         unsigned int result = cobj->getGLVBID();
         ok &= uint32_to_seval(result, &s.rval());
-        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_EditorManager_getGLVBID : Error processing arguments");
+        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_MiddlewareManager_getGLVBID : Error processing arguments");
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 0);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_getGLVBID)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_getGLVBID)
 
-static bool js_cocos2dx_editor_support_EditorManager_getGLIBID(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_getGLIBID(se::State& s)
 {
-    editor::EditorManager* cobj = (editor::EditorManager*)s.nativeThisObject();
-    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_EditorManager_getGLIBID : Invalid Native Object");
+    cocos2d::middleware::MiddlewareManager* cobj = (cocos2d::middleware::MiddlewareManager*)s.nativeThisObject();
+    SE_PRECONDITION2(cobj, false, "js_cocos2dx_editor_support_MiddlewareManager_getGLIBID : Invalid Native Object");
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 0) {
         unsigned int result = cobj->getGLIBID();
         ok &= uint32_to_seval(result, &s.rval());
-        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_EditorManager_getGLIBID : Error processing arguments");
+        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_MiddlewareManager_getGLIBID : Error processing arguments");
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 0);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_getGLIBID)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_getGLIBID)
 
-static bool js_cocos2dx_editor_support_EditorManager_destroyInstance(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_destroyInstance(se::State& s)
 {
     const auto& args = s.args();
     size_t argc = args.size();
     if (argc == 0) {
-        editor::EditorManager::destroyInstance();
+        cocos2d::middleware::MiddlewareManager::destroyInstance();
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 0);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_destroyInstance)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_destroyInstance)
 
-static bool js_cocos2dx_editor_support_EditorManager_getInstance(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_getInstance(se::State& s)
 {
     const auto& args = s.args();
     size_t argc = args.size();
     CC_UNUSED bool ok = true;
     if (argc == 0) {
-        editor::EditorManager* result = editor::EditorManager::getInstance();
-        ok &= native_ptr_to_seval<editor::EditorManager>((editor::EditorManager*)result, &s.rval());
-        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_EditorManager_getInstance : Error processing arguments");
+        cocos2d::middleware::MiddlewareManager* result = cocos2d::middleware::MiddlewareManager::getInstance();
+        ok &= native_ptr_to_seval<cocos2d::middleware::MiddlewareManager>((cocos2d::middleware::MiddlewareManager*)result, &s.rval());
+        SE_PRECONDITION2(ok, false, "js_cocos2dx_editor_support_MiddlewareManager_getInstance : Error processing arguments");
         return true;
     }
     SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 0);
     return false;
 }
-SE_BIND_FUNC(js_cocos2dx_editor_support_EditorManager_getInstance)
+SE_BIND_FUNC(js_cocos2dx_editor_support_MiddlewareManager_getInstance)
 
-SE_DECLARE_FINALIZE_FUNC(js_editor_EditorManager_finalize)
+SE_DECLARE_FINALIZE_FUNC(js_cocos2d_middleware_MiddlewareManager_finalize)
 
-static bool js_cocos2dx_editor_support_EditorManager_constructor(se::State& s)
+static bool js_cocos2dx_editor_support_MiddlewareManager_constructor(se::State& s)
 {
-    editor::EditorManager* cobj = new (std::nothrow) editor::EditorManager();
+    cocos2d::middleware::MiddlewareManager* cobj = new (std::nothrow) cocos2d::middleware::MiddlewareManager();
     s.thisObject()->setPrivateData(cobj);
     se::NonRefNativePtrCreatedByCtorMap::emplace(cobj);
     return true;
 }
-SE_BIND_CTOR(js_cocos2dx_editor_support_EditorManager_constructor, __jsb_editor_EditorManager_class, js_editor_EditorManager_finalize)
+SE_BIND_CTOR(js_cocos2dx_editor_support_MiddlewareManager_constructor, __jsb_cocos2d_middleware_MiddlewareManager_class, js_cocos2d_middleware_MiddlewareManager_finalize)
 
 
 
 
-static bool js_editor_EditorManager_finalize(se::State& s)
+static bool js_cocos2d_middleware_MiddlewareManager_finalize(se::State& s)
 {
-    CCLOGINFO("jsbindings: finalizing JS object %p (editor::EditorManager)", s.nativeThisObject());
+    CCLOGINFO("jsbindings: finalizing JS object %p (cocos2d::middleware::MiddlewareManager)", s.nativeThisObject());
     auto iter = se::NonRefNativePtrCreatedByCtorMap::find(s.nativeThisObject());
     if (iter != se::NonRefNativePtrCreatedByCtorMap::end())
     {
         se::NonRefNativePtrCreatedByCtorMap::erase(iter);
-        editor::EditorManager* cobj = (editor::EditorManager*)s.nativeThisObject();
+        cocos2d::middleware::MiddlewareManager* cobj = (cocos2d::middleware::MiddlewareManager*)s.nativeThisObject();
         delete cobj;
     }
     return true;
 }
-SE_BIND_FINALIZE_FUNC(js_editor_EditorManager_finalize)
+SE_BIND_FINALIZE_FUNC(js_cocos2d_middleware_MiddlewareManager_finalize)
 
-bool js_register_cocos2dx_editor_support_EditorManager(se::Object* obj)
+bool js_register_cocos2dx_editor_support_MiddlewareManager(se::Object* obj)
 {
-    auto cls = se::Class::create("EditorManager", obj, nullptr, _SE(js_cocos2dx_editor_support_EditorManager_constructor));
+    auto cls = se::Class::create("MiddlewareManager", obj, nullptr, _SE(js_cocos2dx_editor_support_MiddlewareManager_constructor));
 
-    cls->defineFunction("addTimer", _SE(js_cocos2dx_editor_support_EditorManager_addTimer));
-    cls->defineFunction("removeTimer", _SE(js_cocos2dx_editor_support_EditorManager_removeTimer));
-    cls->defineFunction("getGLVBID", _SE(js_cocos2dx_editor_support_EditorManager_getGLVBID));
-    cls->defineFunction("getGLIBID", _SE(js_cocos2dx_editor_support_EditorManager_getGLIBID));
-    cls->defineStaticFunction("destroyInstance", _SE(js_cocos2dx_editor_support_EditorManager_destroyInstance));
-    cls->defineStaticFunction("getInstance", _SE(js_cocos2dx_editor_support_EditorManager_getInstance));
-    cls->defineFinalizeFunction(_SE(js_editor_EditorManager_finalize));
+    cls->defineFunction("addTimer", _SE(js_cocos2dx_editor_support_MiddlewareManager_addTimer));
+    cls->defineFunction("removeTimer", _SE(js_cocos2dx_editor_support_MiddlewareManager_removeTimer));
+    cls->defineFunction("getGLVBID", _SE(js_cocos2dx_editor_support_MiddlewareManager_getGLVBID));
+    cls->defineFunction("getGLIBID", _SE(js_cocos2dx_editor_support_MiddlewareManager_getGLIBID));
+    cls->defineStaticFunction("destroyInstance", _SE(js_cocos2dx_editor_support_MiddlewareManager_destroyInstance));
+    cls->defineStaticFunction("getInstance", _SE(js_cocos2dx_editor_support_MiddlewareManager_getInstance));
+    cls->defineFinalizeFunction(_SE(js_cocos2d_middleware_MiddlewareManager_finalize));
     cls->install();
-    JSBClassType::registerClass<editor::EditorManager>(cls);
+    JSBClassType::registerClass<cocos2d::middleware::MiddlewareManager>(cls);
 
-    __jsb_editor_EditorManager_proto = cls->getProto();
-    __jsb_editor_EditorManager_class = cls;
+    __jsb_cocos2d_middleware_MiddlewareManager_proto = cls->getProto();
+    __jsb_cocos2d_middleware_MiddlewareManager_class = cls;
 
     se::ScriptEngine::getInstance()->clearException();
     return true;
@@ -395,17 +395,17 @@ bool register_all_cocos2dx_editor_support(se::Object* obj)
 {
     // Get the ns
     se::Value nsVal;
-    if (!obj->getProperty("jsbEditor", &nsVal))
+    if (!obj->getProperty("middleware", &nsVal))
     {
         se::HandleObject jsobj(se::Object::createPlainObject());
         nsVal.setObject(jsobj);
-        obj->setProperty("jsbEditor", nsVal);
+        obj->setProperty("middleware", nsVal);
     }
     se::Object* ns = nsVal.toObject();
 
-    js_register_cocos2dx_editor_support_EditorManager(ns);
+    js_register_cocos2dx_editor_support_MiddlewareManager(ns);
     js_register_cocos2dx_editor_support_Texture2D(ns);
     return true;
 }
 
-#endif //#if USE_EDITOR_SUPPORT > 0
+#endif //#if USE_MIDDLEWARE > 0
