+++ templates/js-template-link/frameworks/runtime-src/proj.android-studio/app/jni/Application.mk
@@ -21,7 +21,7 @@ USE_NET_WORK := 1
 USE_SPINE := 1
 USE_DRAGONBONES := 1
 USE_TIFF := 1
-USE_EDITOR_SUPPORT := 1
+USE_MIDDLEWARE := 1
 
 APP_CPPFLAGS += -DUSE_GFX_RENDERER=$(USE_GFX_RENDERER)
 APP_CPPFLAGS += -DUSE_VIDEO=${USE_VIDEO}
@@ -31,7 +31,7 @@ APP_CPPFLAGS += -DUSE_NET_WORK=${USE_NET_WORK}
 APP_CPPFLAGS += -DUSE_SPINE=${USE_SPINE}
 APP_CPPFLAGS += -DUSE_DRAGONBONES=${USE_DRAGONBONES}
 APP_CPPFLAGS += -DCC_USE_TIFF=${USE_TIFF}
-APP_CPPFLAGS += -DUSE_EDITOR_SUPPORT=${USE_EDITOR_SUPPORT}
+APP_CPPFLAGS += -DUSE_MIDDLEWARE=${USE_MIDDLEWARE}
 
 USE_ANY_SDK := 1
