+++ tools/tojs/cocos2dx_dragonbones.ini
@@ -1,7 +1,7 @@
 [cocos2dx_dragonbones]
 prefix = cocos2dx_dragonbones
 
-target_namespace = jsbDragonBones
+target_namespace = dragonBones
 
 android_headers = 
 
@@ -20,7 +20,7 @@ cxxgenerator_headers =
 # extra arguments for clang
 extra_arguments = %(android_headers)s %(clang_headers)s %(cxxgenerator_headers)s %(cocos_headers)s %(android_flags)s %(clang_flags)s %(cocos_flags)s %(extra_flags)s 
 
-headers = %(cocosdir)s/cocos/editor-support/dragonBones-creator-support/CCDragonBonesHeaders.h 
+headers = %(cocosdir)s/cocos/editor-support/dragonbones-creator-support/CCDragonBonesHeaders.h 
 
 classes = DragonBonesData CCFactory CCArmatureDisplay Armature CCSlot WorldClock Animation Slot BaseFactory BaseObject TextureData CCTextureData TextureAtlasData CCTextureAtlasData AnimationState EventObject Bone Transform Matrix TransformObject ArmatureData BoneData SlotData SkinData AnimationData
 
@@ -68,7 +68,7 @@ remove_prefix =
 
 classes_have_no_parents = WorldClock Slot AnimationData
 
-base_classes_to_skip = Ref IArmatureProxy IAnimatable IEditor
+base_classes_to_skip = Ref IArmatureProxy IAnimatable IMiddleware
 
 rename_functions = CCArmatureDisplay::[getArmature=armature],
                    CCFactory::[getFactory=getInstance]
