+++ tools/tojs/cocos2dx_editor_support.ini
@@ -1,11 +1,11 @@
 [cocos2dx_editor_support]
 prefix = cocos2dx_editor_support
 
-target_namespace = jsbEditor
+target_namespace = middleware
 
 android_headers = 
 
-macro_judgement = #if USE_EDITOR_SUPPORT > 0
+macro_judgement = #if USE_MIDDLEWARE > 0
 
 android_flags = -target armv7-none-linux-androideabi -D_LIBCPP_DISABLE_VISIBILITY_ANNOTATIONS -DANDROID -D__ANDROID_API__=14 -gcc-toolchain %(gcc_toolchain_dir)s --sysroot=%(androidndkdir)s/platforms/android-14/arch-arm  -idirafter %(androidndkdir)s/sources/android/support/include -idirafter %(androidndkdir)s/sysroot/usr/include -idirafter %(androidndkdir)s/sysroot/usr/include/arm-linux-androideabi -idirafter %(clangllvmdir)s/lib64/clang/5.0/include -I%(androidndkdir)s/sources/cxx-stl/llvm-libc++/include
 
@@ -20,11 +20,11 @@ cxxgenerator_headers =
 # extra arguments for clang
 extra_arguments = %(android_headers)s %(clang_headers)s %(cxxgenerator_headers)s %(cocos_headers)s %(android_flags)s %(clang_flags)s %(cocos_flags)s %(extra_flags)s
 
-headers = %(cocosdir)s/cocos/editor-support/editor-adapter.h %(cocosdir)s/cocos/editor-support/EditorManager.h
+headers = %(cocosdir)s/cocos/editor-support/middleware-adapter.h %(cocosdir)s/cocos/editor-support/MiddlewareManager.h
 
 replace_headers = 
 
-classes = Texture2D EditorManager
+classes = Texture2D MiddlewareManager
 
 classes_need_extend =
