+++ tools/tojs/cocos2dx_spine.ini
@@ -1,7 +1,7 @@
 [cocos2dx_spine]
 prefix = cocos2dx_spine
 
-target_namespace = jsbSpine
+target_namespace = spine
 
 android_headers = 
 
@@ -35,7 +35,7 @@ remove_prefix =
 
 classes_have_no_parents =
 
-base_classes_to_skip = Ref IEditor
+base_classes_to_skip = Ref IMiddleware
 
 abstract_classes =
