+++ tools/tojs/genbindings.py
@@ -144,14 +144,14 @@ def main():
         output_dir = '%s/cocos/scripting/js-bindings/auto' % project_root
 
         cmd_args = {
-                    'cocos2dx.ini' : ('cocos2d-x', 'jsb_cocos2dx_auto'),
-                    'cocos2dx_webview.ini': ('webview', 'jsb_webview_auto'),
-                    'cocos2dx_video.ini': ('video', 'jsb_video_auto'),
-                    'cocos2dx_audioengine.ini' : ('cocos2dx_audioengine', 'jsb_cocos2dx_audioengine_auto'),
-                    'cocos2dx_extension.ini' : ('cocos2dx_extension', 'jsb_cocos2dx_extension_auto'),
-                    'cocos2dx_network.ini' : ('cocos2dx_network', 'jsb_cocos2dx_network_auto'),
-                    'gfx.ini': ('gfx', 'jsb_gfx_auto'),
-                    'renderer.ini': ('renderer', 'jsb_renderer_auto'),
+                    # 'cocos2dx.ini' : ('cocos2d-x', 'jsb_cocos2dx_auto'),
+                    # 'cocos2dx_webview.ini': ('webview', 'jsb_webview_auto'),
+                    # 'cocos2dx_video.ini': ('video', 'jsb_video_auto'),
+                    # 'cocos2dx_audioengine.ini' : ('cocos2dx_audioengine', 'jsb_cocos2dx_audioengine_auto'),
+                    # 'cocos2dx_extension.ini' : ('cocos2dx_extension', 'jsb_cocos2dx_extension_auto'),
+                    # 'cocos2dx_network.ini' : ('cocos2dx_network', 'jsb_cocos2dx_network_auto'),
+                    # 'gfx.ini': ('gfx', 'jsb_gfx_auto'),
+                    # 'renderer.ini': ('renderer', 'jsb_renderer_auto'),
                     'cocos2dx_spine.ini': ('cocos2dx_spine','jsb_cocos2dx_spine_auto'),
                     'cocos2dx_editor_support.ini': ('cocos2dx_editor_support','jsb_cocos2dx_editor_support_auto'),
                     'cocos2dx_dragonbones.ini': ('cocos2dx_dragonbones','jsb_cocos2dx_dragonbones_auto'),
